fix(analytics): keep the explanation's date axis honest, and make admin_only settable
Six findings from the second review round.

`admin_only` existed in the config struct and was enforced, but neither
Helm nor Compose could set it — a stand where people bring their own keys
could not open the route to them. Both now render it, defaulting on.

The prompt clipped `bucket_starts` independently of the series, so a
snapshot with sixty-four dates and a ten-point line told the model the
line was indexed by all sixty-four. Dates that do not index every line
are now dropped rather than sent misaligned: the model can decline to say
when, but it can no longer say the wrong when.

Bucket contributors were deduplicated by display name, so two colleagues
with the same name counted as one — disagreeing with the chart's own
line, which has always keyed on the person. Keyed on the person id now,
names resolved for display.

The trend grid asked for a 34rem minimum track, which is wider than the
content box below about 570px. `min(34rem, 100%)` lets a narrow container
fall back to one full-width track.

`live_tests.rs` reached 570 lines, past the repo's 400-line split rule —
it is a directory now, one module per route group, harness shared through
`super`. Three comments that restated their own code are gone.

Left alone: the generated stand models type optional non-nullable arrays
as nullable. That is how the shared generator has always rendered them —
sixty-eight fields in this one file predate this branch — so it is a
change to `tests/generate_schemas.py` and every service's models, not to
this PR.
