+++ charts/insight/templates/secrets.yaml
@@ -186,6 +186,7 @@ stringData:
   APP__gears__analytics__config__identity_url:        {{ printf "http://%s:8082" (include "insight.identityResolution.host" .) | quote }}
   APP__gears__analytics__config__visibility_policy:   {{ .Values.identityResolution.visibilityPolicy | default "org_chart" | quote }}
   APP__gears__analytics__config__ai_assist__enabled:  {{ .Values.analytics.aiAssist.enabled | default false | quote }}
+  APP__gears__analytics__config__ai_assist__admin_only: {{ .Values.analytics.aiAssist.adminOnly | default true | quote }}
   APP__gears__analytics__config__ai_assist__model:    {{ .Values.analytics.aiAssist.model | default "claude-sonnet-5" | quote }}
   {{- if .Values.analytics.aiAssist.enabled }}
   APP__gears__analytics__config__ai_assist__token_encryption_key: {{ required "analytics.aiAssist.tokenEncryptionKey is required when analytics.aiAssist.enabled is true (openssl rand -base64 32)" .Values.analytics.aiAssist.tokenEncryptionKey | quote }}
